revamp Scorebox, make dynamic, remove old component
