+++ client/src/components/ScoreBox.jsx
@@ -1,57 +1,59 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect} from 'react';
+import { MDBContainer, MDBTabContent, MDBNav, MDBNavItem, MDBNavLink } from "mdbreact";
 import { connect } from 'react-redux';
-import { MDBBadge } from 'mdbreact';
-import moment from 'moment'
-import englishStrings from 'react-timeago/lib/language-strings/en'
-import buildFormatter from 'react-timeago/lib/formatters/buildFormatter'
+import ScoreTabs from './ScoreTabs'
 import Pagination from "./Pagination";
-import DiffTags from './DiffTag'
-import '../styles/ScoreBox.scss';
+import '../styles/ScoreBox.scss'
 
-function ScoreBox(props) {
+function ScoreOverview(props) {
+    let [activeItem, setActiveItem] = useState('1')
     let [allScores, setAllScores] = useState([])
+    let [pageLimit, setPageLimit] = useState(5)
+    let [offset, setOffset] = useState(5)
     let [currentScores, setCurrentScores] = useState([])
-
-    const formatter = buildFormatter(englishStrings)
+    const { data, size } = props;
     const totalScores = allScores.length;
-
-    useEffect(() => {
-        setAllScores(props.data);
-    }, [props.data])
 
     const onPageChanged = data => {
         const { currentPage, pageLimit } = data;
         const offset = (currentPage - 1) * pageLimit;
+        setPageLimit(pageLimit)
+        setOffset((currentPage - 1) * pageLimit)
         setCurrentScores(allScores.slice(offset, offset + pageLimit))
     };
 
+    const toggle = tab => {
+        if (activeItem !== tab) setActiveItem(tab)
+    };
+
+    useEffect(() => {
+        if (activeItem === '1') setAllScores(data.scoresRecent)
+        if (activeItem === '2') setAllScores(data.scoresTop)
+    }, [data, activeItem])
+
+    useEffect(() => {
+        setCurrentScores(allScores.slice(offset, offset + pageLimit))
+    }, [allScores])
+
     if (totalScores === 0) return null;
     else return (
-        <div id="scorebox" className="card-container score-box scores-size-m">
-            <div className="col-md-12">
-                <h3>Latest Scores</h3>
-                <table className="table table-box table-hover">
-                    <thead>
-                        <tr>
-                            <th className="rank" scope="col">Rank</th>
-                            <th className="song" scope="col">Song</th>
-                            <th className="score" scope="col">Score</th>
-                            <th className="time" scope="col">Time</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {currentScores.map((data, index) => {
-                            return(
-                                <tr key={index}>
-                                    <td><MDBBadge color="pink">{data.rank}</MDBBadge></td>
-                                    <td className="song"><DiffTags diff={data.difficulty} /><MDBBadge color="dark">{data.songAuthorName} - {data.songName}</MDBBadge></td>
-                                    <td><MDBBadge color="orange">{data.score}</MDBBadge></td>
-                                    <td><span className="card-link blue-text"><b>{moment(data.timeSet).format('lll')}</b></span></td>
-                                </tr>
-                            )})
-                        }
-                    </tbody>
-                </table>
+        <MDBContainer id="scorebox" className={"card-container score-box scores-size-" + size}>
+                <h3>My Scores</h3>
+                <MDBNav className="nav-tabs mt-5">
+                    <MDBNavItem>
+                        <MDBNavLink link to="#" active={activeItem === '1'} onClick={() => toggle('1')} role="tab" >
+                            Recent
+                        </MDBNavLink>
+                    </MDBNavItem>
+                    <MDBNavItem>
+                        <MDBNavLink link to="#" active={activeItem === '2'} onClick={() => toggle('2')} role="tab" >
+                            Top
+                        </MDBNavLink>
+                    </MDBNavItem>
+                </MDBNav>
+                <MDBTabContent activeItem={activeItem} >
+                    <ScoreTabs tabId={activeItem} data={currentScores} size={size}/>
+                </MDBTabContent>
                 <div className="pagination">
                     <div className="d-flex flex-row py-4 align-items-center">
                     <Pagination
@@ -62,8 +64,7 @@ function ScoreBox(props) {
                     />
                     </div>
                 </div>
-            </div>
-        </div>
+        </MDBContainer>
     )
 }
 
@@ -72,4 +73,5 @@ function mapStateToProps(reduxState){
         userdata : reduxState.userdata
     };
 }
-export default connect(mapStateToProps)(ScoreBox);
+
+export default connect(mapStateToProps)(ScoreOverview);
