+++ client/src/components/ScoreTabs.jsx
@@ -6,13 +6,14 @@ import DiffTags from './DiffTag'
 
 export default function ScoreTabs(props) {
     const { tabId, data, size } = props
+
     return (
-        <MDBTabPane tabId={tabId} role="tabpanel" className="score-box scores-size-l">
+        <MDBTabPane tabId={tabId} role="tabpanel" className={"score-box tabs-size-" + size}>
             <p className="mt-2">
                 <div className="col-md-12">
                     <p>Recent Scores</p>
                     <p>(count: {data.length})</p>
-                    <table className="table table-hover">
+                    <table className="table table-box table-hover">
                         <thead>
                             <tr>
                             <th className="rank" scope="col">Rank</th>
@@ -35,7 +36,6 @@ export default function ScoreTabs(props) {
                                         </td>
                                         <td><MDBBadge color="orange">{data.score}</MDBBadge></td>
                                         <td><b className="card-link blue-text">{moment(data.timeSet).format('lll')}</b></td>
-                                        {/* <td><p>ein ganz langer Text</p></td> */}
                                     </tr>
                                 )})
                             }
