+++ client/src/components/pages/MyProfile.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import api from '../../api';
 import { MDBContainer, MDBRow, MDBCol, MDBBtn, MDBIcon, MDBInput } from 'mdbreact';
-import ScoreOverview from "../ScoreOverview";
+import ScoreBox from "../ScoreBox";
 import '../../styles/pages/MyProfile.scss'
 
 function MyProfile(props) {
@@ -48,7 +48,7 @@ function MyProfile(props) {
                         </MDBCol>
                     </MDBRow>
                 </MDBContainer>
-                {scoreData && <ScoreOverview dataRecent={scoreData.scoresRecent} dataTop={scoreData.scoresTop} />}
+                {scoreData && <ScoreBox data={scoreData} size="lg"/>}
             </div>
         )
     } else {
