+++ client/src/styles/ScoreBox.scss
@@ -1,7 +1,3 @@
-#scorebox {
-  width: 55%;
-}
-
 #page-numbers {
     display: flex;
     list-style: none;
@@ -16,10 +12,6 @@
     padding: 0.5em 0.75rem;
     transition: transform 50ms ease-out;
     margin: 0.4rem;
-  /*  margin-right: 0.3em;
-    color: blue;
-    user-select: none;
-    cursor: pointer;*/
   }
   
   #page-numbers > li:focus,
@@ -38,8 +30,6 @@
     // text-overflow: ellipsis;
   }
   
-
-  
   table img {
     max-width: 40px;
     margin-right: 8px;
@@ -52,3 +42,130 @@
     align-items: center;
     justify-content: center;
   }
+
+#scorebox.card-container {
+  transition: 0.3s;
+  border-radius: 5px;
+  padding: 10px 8px;
+  margin: 12px;
+
+  box-shadow: 
+  0px 0px 1px 3px #fffcdd, 
+  0px 0px 8px 2px #ffeb00, 
+  0px 0px 12px 6px #ffeb00, 
+  
+  inset 0px 0px 6px 2px #ffeb00, 
+  inset 0px 0px 14px 0px #ffeb00, 
+  inset 0px 0px 2px 0px #fffcdd;
+
+  h3 {
+    color: #fffcdd;;
+    text-shadow: 
+    #ffeb00 -2px -1px 5px, 
+    #ffeb00 2px 1px 5px, 
+    #ffeb00 0px 0px 10px, 
+    #ffeb00 0px 0px 15px, 
+    #ffeb00 0px 0px 20px, 
+    #ffeb00 0px 0px 30px;
+    text-align: center;
+  }
+}
+
+
+.score-box {
+  h3{ 
+    color: whitesmoke;
+    font-family: 'NeonTubes2';
+    margin: 5px 0 20px 5px;
+    color: whitesmoke;
+  }
+
+  table {
+    color: whitesmoke;
+  }
+
+  td {
+    vertical-align: middle;
+  }
+
+  th {
+    text-align: left;
+    padding: 0.5rem;
+  }
+
+  .table {
+      td {
+        border-top: none;
+        border-bottom: 1px solid darkgrey;
+        img {
+          margin-left: 1rem;
+        }
+      }
+      tr:last-child {
+        td {
+          border: none
+        }
+      }
+  }
+
+  .table.table td {
+    padding: 0.5rem
+  }
+
+  .card-link {
+    font-size: 0.7rem
+  }
+
+  
+
+  td.song {
+    .badge {
+      margin-left: 2px;
+      text-overflow: ellipsis;
+    }
+  }
+
+  .song-data {
+    display: flex;
+    align-items: center;
+  }
+
+  .card-link {
+    overflow: hidden;
+    white-space: nowrap;
+  }
+}
+
+// === SMALL-SIZE ===
+.scores-size-sm {
+  width: 30%;
+}
+
+// === MEDIUM-SIZE ===
+
+.scores-size-m {
+  width: 55%; 
+  // background-color: lightsalmon;
+
+  th.rank {
+    width: 11%;
+  }
+
+  th.song {
+    width: 53%
+  }
+
+  th.score {
+    width: 17%;
+  }
+
+  th.time {
+    width: 19%;
+  }
+}
+
+// === LARGE-SIZE ===
+// No width limitation
+
+.scores-size-lg {
+}
