--- client/src/styles/ScoreOverview.scss
@@ -1,97 +1,0 @@
-.score-box {
-  h3{ 
-    color: whitesmoke;
-    font-family: 'NeonTubes2';
-    margin: 5px 0 20px 5px;
-    color: whitesmoke;
-  }
-
-  table {
-    color: whitesmoke;
-  }
-
-  td {
-    vertical-align: middle;
-  }
-
-  th {
-    text-align: left;
-    padding: 0.5rem;
-  }
-
-  .table {
-      td {
-        border-top: none;
-        border-bottom: 1px solid darkgrey;
-        img {
-          margin-left: 1rem;
-        }
-      }
-      tr:last-child {
-        td {
-          border: none
-        }
-      }
-  }
-
-  .table.table td {
-    padding: 0.5rem
-  }
-
-  .card-link {
-    font-size: 0.7rem
-  }
-
-  td.song {
-    .badge {
-      margin-left: 2px;
-      text-overflow: ellipsis;
-    }
-  }
-
-  .song-data {
-    display: flex;
-    align-items: center;
-  }
-
-  .card-link {
-    overflow: hidden;
-    white-space: nowrap;
-  }
-}
-
-.scores-size-m {
-  th.rank {
-    width: 11%;
-  }
-
-  th.song {
-    width: 53%
-  }
-
-  th.score {
-    width: 17%;
-  }
-
-  th.time {
-    width: 19%;
-  }
-}
-
-.scores-size-l {
-  // th.rank {
-  //   width: 11%;
-  // }
-
-  // th.song {
-  //   width: 23%
-  // }
-
-  // th.score {
-  //   width: 17%;
-  // }
-
-  th.time {
-    width: 27%;
-  }
-}
